src/shifts/pyproject.toml
[tool.poetry]
name = "shifts"
version = "0.1.0"
description = "Simple REST API application using FastAPI"
authors = ["Mikio Oba <hello@mikio.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.17.6"
fastapi = "^0.78.0"
pydantic = "^1.9.1"
SQLAlchemy = "^1.4.37"
alembic = "^1.8.0"
tenacity = "^8.0.1"
requests = "^2.28.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
psycopg2 = "^2.9.3"
gunicorn = "^20.1.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
black = "^22.3.0"
pylint = "^2.14.2"
requests = "^2.28.0"
python-dotenv = "^0.20.0"
pytest-docker-compose = "^3.2.1"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.18.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
